requirements.txt
# AWS
boto3==1.40.50
botocore==1.40.50
# Speeds up IAM logic
cached-property==2.0.1
# Command line
click==8.3.0
click_option_group==0.5.9
# Templating - required for exclusions YAML files and inserting content to HTML report
jinja2==3.1.6
pyyaml==6.0.3
# We render Markdown glossary files as HTML in the Cloudsplaining report
markdown==3.9
# AWS IAM Logic
policy-sentry==0.14.1
# Schema validation
schema==0.7.7
